PP-10701: Move `EventEntity` to another package
For consistency with the rest of the codebase
