+++ .secrets.baseline
@@ -128,13 +128,13 @@
         "line_number": 1637
       }
     ],
-    "src/main/java/uk/gov/pay/ledger/event/model/EventEntity.java": [
+    "src/main/java/uk/gov/pay/ledger/event/dao/entity/EventEntity.java": [
       {
         "type": "Hex High Entropy String",
-        "filename": "src/main/java/uk/gov/pay/ledger/event/model/EventEntity.java",
+        "filename": "src/main/java/uk/gov/pay/ledger/event/dao/entity/EventEntity.java",
         "hashed_secret": "cbbaa15cf3814df641e9ee9cb279046f93f322eb",
         "is_verified": false,
-        "line_number": 20
+        "line_number": 21
       }
     ],
     "src/test/java/uk/gov/pay/ledger/pact/ContractTest.java": [
@@ -156,5 +156,5 @@
       }
     ]
   },
-  "generated_at": "2023-03-09T16:21:02Z"
+  "generated_at": "2023-03-09T16:37:02Z"
 }
