+++ src/main/java/uk/gov/pay/ledger/agreement/dao/AgreementDao.java
@@ -7,7 +7,7 @@ import org.jdbi.v3.core.statement.Query;
 import uk.gov.pay.ledger.agreement.entity.AgreementEntity;
 import uk.gov.pay.ledger.agreement.resource.AgreementSearchParams;
 import uk.gov.pay.ledger.event.dao.mapper.EventMapper;
-import uk.gov.pay.ledger.event.model.EventEntity;
+import uk.gov.pay.ledger.event.dao.entity.EventEntity;
 
 import java.util.List;
 import java.util.Optional;
