+++ src/main/java/uk/gov/pay/ledger/event/dao/mapper/EventMapper.java
@@ -2,7 +2,7 @@ package uk.gov.pay.ledger.event.dao.mapper;
 
 import org.jdbi.v3.core.mapper.RowMapper;
 import org.jdbi.v3.core.statement.StatementContext;
-import uk.gov.pay.ledger.event.model.EventEntity;
+import uk.gov.pay.ledger.event.dao.entity.EventEntity;
 import uk.gov.pay.ledger.event.model.ResourceType;
 
 import java.sql.ResultSet;
