+++ src/main/java/uk/gov/pay/ledger/queue/eventprocessor/AgreementEventProcessor.java
@@ -2,7 +2,7 @@ package uk.gov.pay.ledger.queue.eventprocessor;
 
 import com.google.inject.Inject;
 import uk.gov.pay.ledger.agreement.service.AgreementService;
-import uk.gov.pay.ledger.event.model.EventEntity;
+import uk.gov.pay.ledger.event.dao.entity.EventEntity;
 import uk.gov.pay.ledger.event.service.EventService;
 
 public class AgreementEventProcessor extends EventProcessor {
