+++ src/main/java/uk/gov/pay/ledger/queue/eventprocessor/EventProcessor.java
@@ -1,6 +1,6 @@
 package uk.gov.pay.ledger.queue.eventprocessor;
 
-import uk.gov.pay.ledger.event.model.EventEntity;
+import uk.gov.pay.ledger.event.dao.entity.EventEntity;
 
 public abstract class EventProcessor {
     public abstract void process(EventEntity event, boolean isANewEvent);
