+++ src/main/java/uk/gov/pay/ledger/transactionsummary/service/TransactionSummaryService.java
@@ -1,7 +1,7 @@
 package uk.gov.pay.ledger.transactionsummary.service;
 
 import com.google.inject.Inject;
-import uk.gov.pay.ledger.event.model.EventEntity;
+import uk.gov.pay.ledger.event.dao.entity.EventEntity;
 import uk.gov.pay.ledger.event.model.SalientEventType;
 import uk.gov.pay.ledger.transaction.entity.TransactionEntity;
 import uk.gov.pay.ledger.transaction.model.TransactionType;
